requirements.txt
Flask
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
redis[hiredis]
psycopg2-binary
alembic
flask-migrate
bcrypt
Flask-WTF
WTForms
email-validator
flask-login
flask-socketio
gunicorn==20.1.0
